Advanced the state of the libbinary for mach-o files. Implemented a MemoryMap so we handle memory accesses safely. Adds the structure for the not yet implemented libraries. Updates the ARM disassembler generator.

+++ src/utilities/Utilities.h
@@ -34,7 +34,7 @@ static inline std::string exec_get_output(std::string cmd) {
 	return result;
 }
 
-static inline std::string integer_to_string(unsigned val, bool hexa = true) {
+static inline std::string integer_to_string(unsigned long long val, bool hexa = true) {
 	std::stringstream ss;
 	if (hexa) {
 		ss << std::hex << "0x";
